+++ package.json
@@ -9,6 +9,7 @@
     "lint": "next lint"
   },
   "dependencies": {
+    "@googleapis/youtube": "^14.0.0",
     "@supabase/supabase-js": "^2.39.8",
     "@tanstack/react-query": "^5.28.4",
     "jotai": "^2.7.1",
