+++ src/app/page.tsx
@@ -1,3 +1,9 @@
+import Footer from "./sherd/Footer";
+
 export default function Home() {
-  return <div></div>;
+  return (
+    <div>
+      <Footer />
+    </div>
+  );
 }
