+++ src/app/sherd/reset.css
@@ -107,6 +107,8 @@ section {
 }
 body {
   line-height: 1;
+  height: 900px;
+  width: 1920px;
 }
 ol,
 ul {
